// shared/types/browser.ts
export interface BaseMessage {
  type: string
  data?: any
  timestamp?: number
}

export interface TabChangeInfo {
  audible?: boolean;
  discarded?: boolean;
  favIconUrl?: string;
  mutedInfo?: Browser.tabs.MutedInfo;
  pinned?: boolean;
  status?: 'loading' | 'complete';
  title?: string;
  url?: string;
}

export interface ActiveTab {
  tabId: number
  pageId: number
  url: string
  domain: string
  isVisible: boolean
  isIdle: boolean
  lastActivityTime: number
}